REST: add explicit cast to sitemapSize calcuation to avoid warning
Bug: T411580
(cherry picked from commit 3e3755c4f58bc54eb96c55f7152f131850822fa2)

+++ includes/Rest/Handler/SitemapHandlerBase.php
@@ -32,7 +32,7 @@ abstract class SitemapHandlerBase extends Handler {
 		$this->enabled = $apiConf['enabled'] ?? false;
 		$this->indexSize = $apiConf['sitemapsPerIndex'] ?? 50_000;
 		$variants = SitemapGenerator::getVariants( $contLang, $languageConverterFactory );
-		$this->sitemapSize = ( $apiConf['pagesPerSitemap'] ?? 10_000 ) / ( count( $variants ) + 1 );
+		$this->sitemapSize = (int)( ( $apiConf['pagesPerSitemap'] ?? 10_000 ) / ( count( $variants ) + 1 ) );
 		$this->expiry = $apiConf['expiry'] ?? 3600;
 	}
